add react app creator

+++ mirage/fileable.py
@@ -43,7 +43,8 @@ def mkdir(path):
 
 
 def rm(path):
-    shutil.rmtree(path)
+    if os.path.isdir(path): shutil.rmtree(path)
+    if os.path.isfile(path): os.remove(path)
 
 
 def cwd():
